// src/app/shared/interfaces/ISearch.interface.ts
export interface ISearch {
    _id: string;
    title: string;
    applylink: string;
    jd: string;
    companyname: string;
    location: string;
    experience: string;
    salary: string;
    type: string;
    skills: string;
    startdate: string;
    enddate: string;
    created: string;
    source: string;
    timestamp: number;
    __v: 0;
}